// src/api/models/IdentityProviderRepresentation.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IdentityProviderRepresentation = {
    addReadTokenRoleOnCreate?: boolean;
    alias?: string;
    config?: Record<string, any>;
    displayName?: string;
    enabled?: boolean;
    firstBrokerLoginFlowAlias?: string;
    internalId?: string;
    linkOnly?: boolean;
    postBrokerLoginFlowAlias?: string;
    providerId?: string;
    storeToken?: boolean;
    trustEmail?: boolean;
};